Add files via upload
First working version of the script for database generation.
Takes as an input the path to a directory, which should contain a folder for each speaker. 
SPL difference (attenuation to one of the speech audios) can be selected, however there is no measurement done beforehand so unless both audios have been normalized to the same level, the difference may not be as specified.